Use unsigned int instead of unsigned long
I was getting errors when trying to use unsigned long in OpenCL Kernels
in windows (it worked in Linux).  So for now I am changing the kernels
to use unsigned int to represent size.  Long term I need to understand
better what types are and are not supported in OpenCL (and if it differs
between implementations of OpenCL).